+++ resolwe_bio/tests/workflows/test_cut_and_run.py
@@ -40,6 +40,10 @@ class CutAndRunTestCase(BioProcessTestCase):
             "options_sieve": {
                 "max_frag_length": 120,
             },
+            "options_misc": {
+                "bw_binsize": 50,
+                "bw_timeout": 30,
+            },
         }
 
         self.run_process("workflow-cutnrun", input_workflow)